% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beetles.R
\docType{data}
\name{Beetles}
\alias{Beetles}
\title{Beetle Diversity Data}
\format{A \code{data.frame} with 96 rows and 582 variables}
\source{
TO Crist, JA Veech, JC Gering, and KS Summerville. 2003. Partitioning
    species diversity across landscapes and regions: A hierarchical analysis
    of alpha, beta, and gamma diversity. The American Naturalist 162:734-743.
}
\usage{
Beetles
}
\description{
A data set containing canopy beetle abundance data for 582 species collected
from 96 trees across 24 stands across 6 sites across 2 ecoregions. 'ECO' is
ecoregion, 'SITE' is site, 'Hab_St' is stand, and 'TREE' is tree.
}
\keyword{datasets}
