% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions.R
\name{partiplot}
\alias{partiplot}
\title{Plotting Partitioned Diversity}
\usage{
partiplot(part.obj, beta.type = "add", plot.type = "bar")
}
\arguments{
\item{part.obj}{an object of class \code{"partition"}, a result of a call to
\code{\link{partition}}}

\item{beta.type}{the method of partition to be plotted; if \code{"add"}, additive diversity will be plotted;
if \code{"mult"}, multiplicative diversity will be plotted.}

\item{plot.type}{the type of graph; if \code{"bar"}, a stacked bar chart will be plotted.
if \code{"line"}, a line plot will be plotted with 95\% CI error bars.}
}
\value{
A \code{\link{ggplot}} graphic of the observed and expected partition.
    \code{"bar"} represents a stacked bar chart with the y-axis representing a
    proportion of the gamma (regional) diversity explained by each level. \code{"line"}
    represents a line plot of the change in alpha- and beta-diversity of both observed
    and expected data. Error bars represent 95\% CIs of null statistical distribution.
}
\description{
Graphical display of partitioned data. Graphics include a stacked bar chart
    and line plot. Both observed partition and expected partition (i.e., mean of the null
    statistical distributions) are shown in both the bar chart and line plot.
}
\details{
Function \code{partiplot} visualizes the partitioned data from an object of
    class \code{partition}.

The function uses \code{\link[ggplot2]{ggplot}} to visualize the partitioned data.
    To save the graphic, use the \code{\link[ggplot2]{ggsave}]} function.
}
\examples{
\dontrun{
## Plot a stacked bar chart of the additive partitioned data
partiplot(part.obj, beta.type = "add", plot.type = "bar")

## Plot a line plot of the multiplicative partitioned data
partiplot(part.obj, beta.type = "mult", plot.type = "line")
}
}
\seealso{
The diversiting partitioning function \code{\link{partition}}.
}
