% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionFunctions.R
\name{partition}
\alias{partition}
\alias{print.partition}
\title{Partitioning Diversity}
\usage{
partition(sp, h.level, low.level = 1, q = 0, hyp.test = "INDIVIDUAL",
  sim.rand = 1000)

\method{print}{partition}(x, ...)
}
\arguments{
\item{sp}{Data frame containing a species matrix that includes variables
coding for the different scales by which diversity is to be partitioned.
The columns containing levels by which diversity is to be partitioned
\strong{must} be of class \code{factor}.}

\item{h.level}{Vector or list containing column names coding for the
different scales by which diversity is to be partitioned. These must be
in increasing scale.}

\item{low.level}{Integer representing the lowest level of diversity to be
partitioned. This is the lowest ecologically relevant level. Defaults to
\code{1}. Must be \code{>1} when \code{hyp.test = "SAMPLE"}.}

\item{q}{Integer representing Hill Number q-diversity metrics. \code{0}
represents species richness; \code{1} represents Shannon-diversity; and
\code{2} represents Simpsons-diversity. See Jost (2007) for more
information.}

\item{hyp.test}{Method of hypothesis testing to be used; for hypothesis
testing: \code{hyp.test = "INDIVIDUAL"} and \code{hyp.test = "SAMPLE"};
for calculation of observed values only: \code{hyp.test = "NONE"}.}

\item{sim.rand}{Integer representing the number of randomizations to be run
for hypothesis testing; defaults to \code{1000}. If
\code{hyp.test = "NONE"}, \code{sim.rand} can be ignored.}

\item{x}{object returned by \code{partition} of class \code{"partition"}}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of class \code{"partition"}. This object is a list of data
    frames. Calling an object of class \code{"partition"} will print the
    \code{$Div} and \code{$Hyp} data frames; these data frames contain the
    partitioned data and p-values from significance tests, respectively. For
    more information from the object and interpretations of significance
    tests, please use the \code{\link{summary.partition}} function.

\item{\code{$Div}}{observed partitioned diversity}

\item{\code{$Hyp}}{p-values from significance testing}

\item{\code{$Rand.Beta.Add}}{expected additive Beta diversity}

\item{\code{$Rand.Beta.Mult}}{expected multipicative Beta diversity}

\item{\code{$Rand.Alpha}}{expected alpha diversity at the lowest
    ecologically important level (set by \code{low.level})}

\item{\code{$Test}}{\code{"INDIVIDUAL"}, \code{"SAMPLE"}, or
    \code{"NONE"} - set by \code{hyp.test}, passed to object to be used in
    support and generic functions}

\item{\code{$q}}{integer set by \code{q}, passed to object to be used
    in support and generic functions}

\item{\code{$Randomizations}}{integer set by \code{sim.rand}, passed
    to object to be used in support and generic functions}
}
\description{
\code{partition} is used to calculate alpha, beta, and gamma
    diversity of both balanced and unbalanced designs. It can be used to
    carry out diversity partitioning at both single and multiple scales, as
    well as nested designs. \code{partition} can run hypothesis testing based
    on distributions generated from two randomizations methods (see below).
}
\details{
Diversity partitioning is a method of decomposing a total amount of
    diversity (\emph{gamma} - \eqn{\gamma}) into the components of mean
    diversity within samples (\emph{alpha} - \eqn{\alpha}) and diversity
    among samples (\emph{beta} - \eqn{\beta}). It can be used with a wide
    variety of diversity metrics, specifically the Hill Number
    \emph{q}-diversity metrics (Jost 2007). \eqn{\gamma} and \eqn{\alpha} are
    calculated using equations 3 - 6 of Chao et al. (2012), with equations 5
    and 6 used when \code{q = 1} and equations 3 and 4 used in all other
    cases.

The \code{partition} function can be applied to a variety of data
    sets, including those with an unbalanced sampling design, substantial
    variation in the number of individuals within those samples, and
    hierarchical sampling designs (multiple nested levels).

\code{partition} calculates alpha and beta diversity and uses
    randomization to derive expected values of alpha and beta diversity that
    would be obtained if individuals (\code{INDIVIDUAL}) or samples
    (\code{SAMPLE}) were randomly distributed. This randomization allows for
    significance testing of the observed diversity estimates. The statistical
     rationale and operational description of \code{INDIVIDUAL}- and
    \code{SAMPLE}-based randomization can be found in Crist et al. (2003).

The \code{partition} function is the R equivalent of the PARTITION
    software developed by Crist et al. (2003).

At the highest sampling level (h), the diversity components are
    calculated as follows:
    \deqn{Additive: \beta (h) = \gamma - \alpha (h)}
    \deqn{Multiplicative: \beta (h) = \gamma / \alpha (h)}
    For the lower sampling levels calculated as follows:
    \deqn{Additive: \beta (i) = \alpha (i+1) - \alpha (i)}
    \deqn{Multiplicative: \beta (i) = \alpha (i+1) / \alpha (i)}
}
\examples{
\dontrun{
part.obj <- partition(sp = spiders.spp,
                      h.level = c("SAMPLE", "TREESP"),
                      low.level = 1,
                      q = 0,
                      hyp.test = "INDIVIDUAL",
                      sim.rand = 1000)

print(part.obj)
}
}
