% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARTITIONR.R
\docType{package}
\name{PARTITIONR}
\alias{PARTITIONR}
\alias{PARTITIONR-package}
\title{PARTITIONR: A package for partitioning diversity across spatial scales}
\description{
The PARTITIONR package provides tools for partitioning diversity across spatial scales. It has
    three types of functions.
}
\section{Partition Function}{

The partition function partitions diversity across nested scales.
}

\section{Support Functions}{

The support functions provide an opportunity to visualize the data and output of
    the \code{\link{partition}} function. These functions are good.
}

\section{Community Analysis Functions}{

The community analysis functions are NOT FUNCTIONAL YET.
}

