% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericFunctions.R
\name{summary.partition}
\alias{summary.partition}
\alias{print.summary_partition}
\title{Summarizing Partitioned Diversity}
\usage{
\method{summary}{partition}(object, p.value = "one-sided", ...)

\method{print}{summary_partition}(x, ...)
}
\arguments{
\item{object}{an object of class \code{"partition"}, a result of a call to
\code{\link{partition}}}

\item{p.value}{a character; if \code{"one-sided"}, the p-values are interpreted at < observed and significance at p = 0.05.
if \code{"two-sided"}, the p-values are interpreted at both > and < observed and significance at p = 0.025.}

\item{...}{additional arguments affecting the summary produced}

\item{x}{an obect of class \code{"summary_parition"}, a result of a call to
\code{\link{summary.partition}}}
}
\value{
The observed and mean expected values of alpha and beta diversity, as
    well as the the p-values (number of randomizations < observed diversity)
    of each diversity level and interpretations of these. Values are rounded to number of places given by \eqn{1/# Simulations}

\item{Test}{\code{"INDIVIDUAL"} or \code{"SAMPLE"} randomization as specified in \code{partition} function}

\item{P-value}{\code{"one-sided"} or \code{"two-sided"} interpretation of results}

\item{Randomizations}{number of randomizations run as specified in \code{partition} function}

\item{q}{Hill number as specified in \code{partition} function}

\item{Gamma}{observed gammma (regional) diversity}

\item{Observed}{observed lowest level alpha and additive and multiplicative
    beta diversities from the species matrix supplied to \code{\link{partition}}.}

\item{Expected}{mean lowest level alpha and addtive and multiplicative
    beta diversities from randomizations of \code{\link{partition}}.}

\item{P(<Obs)}{probability of an expected value being lower than the observed value based on number of randomizations}

Significance interpretations depend on \code{p.value}: if \code{"one-sided"}, the p-values are interpreted at
    < observed and significance at p = 0.05. If \code{"two-sided"}, the p-values are interpreted at
    > and < observed and significance at p = 0.025; with 0.975 being significant probability of observed being less than expected.
}
\description{
\code{summary} method for class \code{"partition"}
}
\details{
\code{summary.partition} tries to be smart about formating the
    observed and expected diversity values and significance.

P-values are calculated as \eqn{1 - ((# expected values < observed
    values)/# Simulations)}.

Errors will be thrown if \code{p.value} is not equal to either
    \code{"one-sided"} or \code{"two-sided"}.
}
\examples{
\dontrun{
## One-tailed p-values
summary(partition.obj)

## Two-tailed p-values
summary(partition.obj, p.value = "two-sided")
}
}
\seealso{
The diversiting partitioning function \code{\link{partition}}.
}
